requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
gunicorn
python-dotenv